game-core/movement fix movement generator part2

+++ game-core/src/main/java/com/github/azeroth/game/miscellaneous/Formulas.java
@@ -44,7 +44,7 @@ public class Formulas {
                 }
             }
 
-            level = (int) ((pl_level - 7) - (count - 1));
+            level = (pl_level - 7) - (count - 1);
         } else {
             level = pl_level - 10;
         }
